refactor(checker): split keyof source display helper
